% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_selection.R
\name{input_importance}
\alias{input_importance}
\title{Determines the importance of each input}
\usage{
input_importance(
  X,
  y,
  q,
  n_init,
  inf_crit = "BIC",
  task = "regression",
  unif = 3,
  addition = FALSE,
  X_full = NULL,
  maxit = 1000,
  ...
)
}
\arguments{
\item{X}{Matrix of covariates}

\item{y}{Vector of response}

\item{q}{Number of hidden nodes}

\item{n_init}{Number of random initialisations (tracks)}

\item{inf_crit}{Information criterion: \code{"BIC"} (default), \code{"AIC"} or
\code{"AICc"}}

\item{task}{\code{"regression"} (default) or \code{"classification"}}

\item{unif}{Random initial values max value}

\item{addition}{Switch for addition step (default FALSE)}

\item{X_full}{Full matrix of covariates if X has some dropped}

\item{maxit}{maximum number of iterations for nnet (default = 100)}

\item{...}{additional argument for nnet}
}
\value{
The least important input node
}
\description{
Calculates the importance of each input model based on information criterion
and returns which node is least important
}
