% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_selection.R
\name{input_node_sel}
\alias{input_node_sel}
\title{Selects input layer structure}
\usage{
input_node_sel(
  X,
  y,
  q,
  n_init,
  type = "bulk",
  inf_crit = "BIC",
  unif = 3,
  maxit = 1000,
  ...
)
}
\arguments{
\item{X}{Matrix of covariates}

\item{y}{Vector of response}

\item{q}{Number of hidden nodes}

\item{n_init}{Number of random initialisations (tracks)}

\item{type}{Selection type: \code{"bulk"} (the default) or \code{"step"}}

\item{inf_crit}{Information criterion: \code{"BIC"} (default), \code{"AIC"} or
\code{"AICc"}}

\item{unif}{Random initial values max value}

\item{maxit}{maximum number of iterations for nnet (default = 100)}

\item{...}{additional argument for nnet}
}
\value{
Inputs dropped from model
}
\description{
Performs either bulk or step-wise input node selection.
}
