% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_likelihood}
\alias{log_likelihood}
\title{Calculates normal log-likelihood of neural network}
\usage{
log_likelihood(W, X, Y, q)
}
\arguments{
\item{W}{Weight vector}

\item{X}{Input data}

\item{Y}{Output data}

\item{q}{Number of hidden units}
}
\value{
log-likelihood value
}
\description{
Calculates normal log-likelihood of neural network
}
