% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{my_runif}
\alias{my_runif}
\title{Gernerate uniformly random variables in interval not containing values close
to zero#'}
\usage{
my_runif(n = 1, unif = 1, lower = 0.5)
}
\arguments{
\item{n}{number of values}

\item{unif}{max value}

\item{lower}{min value}
}
\value{
vector of values
}
\description{
Gernerate uniformly random variables in interval not containing values close
to zero#'
}
