% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{n_params}
\alias{n_params}
\title{Number of weights in NN}
\usage{
n_params(p, q)
}
\arguments{
\item{p}{number of inputs}

\item{q}{structure of hidden layers}
}
\value{
number of parameters
}
\description{
Number of weights in NN
}
