% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_functions.R
\name{nn_fit_tracks}
\alias{nn_fit_tracks}
\title{Fits various tracks (different random starting values) and chooses best model}
\usage{
nn_fit_tracks(
  X,
  y,
  q,
  n_init,
  inf_crit = "BIC",
  task = "regression",
  unif = 3,
  maxit = 1000,
  ...
)
}
\arguments{
\item{X}{Matrix of covariates}

\item{y}{Vector of response}

\item{q}{Number of hidden nodes}

\item{n_init}{Number of random initialisations (tracks)}

\item{inf_crit}{Information criterion: \code{"BIC"} (default), \code{"AIC"} or
\code{"AICc"}}

\item{task}{\code{"regression"} (default) or \code{"classification"}}

\item{unif}{Random initial values max value}

\item{maxit}{maximum number of iterations for nnet (default = 100)}

\item{...}{additional argument for nnet}
}
\value{
The best model from the different tracks
}
\description{
Fits n_init tracks with different initial values and decides on best model
based on information criteria.
}
