% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_model_sel.R
\name{nn_model_sel}
\alias{nn_model_sel}
\title{Full model selection (input and hidden layer)}
\usage{
nn_model_sel(X, y, Q, n_init, inf_crit = "BIC", unif = 3, maxit = 1000, ...)
}
\arguments{
\item{X}{Matrix of covariates}

\item{y}{Vector of response}

\item{Q}{Candidate number of hidden nodes}

\item{n_init}{Number of random initialisations (tracks)`}

\item{inf_crit}{Information criterion: \code{"BIC"} (default), \code{"AIC"} or
\code{"AICc"}}

\item{unif}{Random initial values max value}

\item{maxit}{maximum number of iterations for nnet (default = 100)}

\item{...}{additional argument for nnet}
}
\value{
Optimal number of hidden nodes
}
\description{
Performs model selection procedure.
}
