% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_pred.R
\name{nn_pred}
\alias{nn_pred}
\title{Network Output Function}
\usage{
nn_pred(X, W, q)
}
\arguments{
\item{X}{Data}

\item{W}{Weight vector}

\item{q}{Number of hidden units}
}
\value{
Prediction for given inputs
}
\description{
Network Output Function
}
