% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_unit}
\alias{remove_unit}
\title{Selects which hidden unit to remove based off of BIC}
\usage{
remove_unit(W, dataX, Y, q, inf_crit = "BIC")
}
\arguments{
\item{W}{Weight vector}

\item{dataX}{Matrix of inputs}

\item{Y}{Output vector}

\item{q}{Number of hidden nodes in model}

\item{inf_crit}{Information criterion}
}
\value{
The node to remove
}
\description{
This function calculates the influence after of each hidden unit in the model
and determines which unit has the least influence on the output
}
