% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectnn.R
\name{selectnn}
\alias{selectnn}
\alias{selectnn.default}
\alias{selectnn.formula}
\title{Neural network model selection}
\usage{
selectnn(...)

\method{selectnn}{default}(
  X,
  y,
  Q,
  n_init,
  inf_crit = "BIC",
  task = "regression",
  unif = 3,
  maxit = 1000,
  ...
)

\method{selectnn}{formula}(formula, data, ...)
}
\arguments{
\item{...}{arguments passed to or from other methods}

\item{X}{Matrix of covariates}

\item{y}{Vector of response}

\item{Q}{Candidate number of hidden nodes}

\item{n_init}{Number of random initialisations (tracks)`}

\item{inf_crit}{Information criterion: \code{"BIC"} (default), \code{"AIC"} or
\code{"AICc"}}

\item{task}{\code{"regression"} (default) or \code{"classification"}}

\item{unif}{Random initial values max value}

\item{maxit}{maximum number of iterations for nnet (default = 100)}

\item{formula}{A formula of the form: response ~ x1 + x2 + ...}

\item{data}{Data frame from which variables specified in formula are to be
taken}
}
\value{
A list with information of the optimal model.
\itemize{
\item \code{nn_hidden} - list of hidden node selection results.
\item \code{nn_input} - list of input node selection results.
\item \code{n_rep_h} - number of hidden node selection steps.
\item \code{n_rep_i} - number of input node selection steps.
\item \code{X} - matrix of the important covariates found.
\item \code{X_full} - matrix of all covariates.
\item \code{dropped} - vector of unimportant covariates.
\item \code{hidden_size} - vector of hidden layer size found at each step.
}
}
\description{
Performs both input and hidden layer selection for neural networks.
}
