% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{weight_recover}
\alias{weight_recover}
\title{Function to rearrange weights due to weight space symmetries}
\usage{
weight_recover(W_true, W_pred, p, q)
}
\description{
Function to rearrange weights due to weight space symmetries
}
