% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden_node_sel.R
\name{hidden_node_sel}
\alias{hidden_node_sel}
\title{Hidden Node Selection}
\usage{
hidden_node_sel(
  X,
  y,
  Q,
  n_init,
  type = "bulk",
  inf_crit = "BIC",
  unif = 3,
  ...
)
}
\arguments{
\item{X}{Matrix of covariates}

\item{y}{Vector of response}

\item{Q}{Possible number of hidden nodes}

\item{n_init}{Number of random initialisations (tracks)}

\item{type}{bulk or step}

\item{inf_crit}{Information criterion}

\item{unif}{Random initial values max value}

\item{...}{additional argument for nnet}
}
\value{
Optimal number of hidden nodes
}
\description{
Performs either bulk or step-wise hidden node selection
}
