% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_importance.R
\name{input_importance}
\alias{input_importance}
\title{Determines the importance of each input}
\usage{
input_importance(X, y, q, n_init, inf_crit = "BIC", unif = 3, ...)
}
\arguments{
\item{X}{Matrix of covarites}

\item{y}{Vector of response}

\item{q}{Number of hidden nodes}

\item{n_init}{Number of random initialisations (tracks)}

\item{inf_crit}{Information criterion}

\item{unif}{Random initial values max value}
}
\value{
The least important input node
}
\description{
Calculates the importance of each input model based on inf_crit and returns
which node is least important
}
