% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_node_sel.R
\name{input_node_sel}
\alias{input_node_sel}
\title{Input Node Selection}
\usage{
input_node_sel(X, y, q, n_init, type = "bulk", inf_crit = "BIC", unif = 3, ...)
}
\arguments{
\item{X}{Matrix of covarites}

\item{y}{Vector of response}

\item{q}{Number of hidden nodes}

\item{n_init}{Number of random initialisations (tracks)}

\item{type}{bulk or step}

\item{inf_crit}{Information criterion}

\item{unif}{Random initial values max value}

\item{...}{additional argument for nnet}
}
\value{
Inputs dropped from model
}
\description{
Performs either bulk or stepwise input node selection
}
