% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_functions.R
\name{nn_pred}
\alias{nn_pred}
\title{Neural network prediction}
\usage{
nn_pred(X, W, q)
}
\arguments{
\item{X}{Data}

\item{W}{Weight vector}

\item{q}{Number of hidden nodes}
}
\value{
Prediction for given inputs
}
\description{
Neural network prediction
}
